게시물 수정&삭제시 권한체크

+++ src/main/java/com/sbs/example/lolHi/controller/usr/ArticleController.java
@@ -72,6 +72,12 @@ public class ArticleController {
 			model.addAttribute("replaceUri", "/usr/member/login");
 			return "common/redirect";
 		}
+		Article article = articleService.getArticleById(id);
+		if (article.getMemberId() != loginedMemberId) {
+			model.addAttribute("msg"," 권한이 없습니다.");
+			model.addAttribute("historyBack", true);
+			return "common/redirect";
+		}
 
 		articleService.deleteArticleById(id);
 
@@ -95,7 +101,11 @@ public class ArticleController {
 		}
 
 		Article article = articleService.getArticleById(id);
-
+		if (article.getMemberId() != loginedMemberId) {
+			model.addAttribute("msg"," 권한이 없습니다.");
+			model.addAttribute("historyBack", true);
+			return "common/redirect";
+		}
 		model.addAttribute("article", article);
 
 		return "usr/article/modify";
@@ -113,6 +123,12 @@ public class ArticleController {
 		if (loginedMemberId == 0) {
 			model.addAttribute("msg", "로그인 후 이용해주세요.");
 			model.addAttribute("replaceUri", "/usr/member/login");
+			return "common/redirect";
+		}
+		Article article = articleService.getArticleById(id);
+		if (article.getMemberId() != loginedMemberId) {
+			model.addAttribute("msg"," 권한이 없습니다.");
+			model.addAttribute("historyBack", true);
 			return "common/redirect";
 		}
